dummyjson-api-test/requirements.txt
requests
pytest
allure-pytest
pyyaml